refactor: update method signatures for epoch and slot to use int type for consistency

+++ extensions/blockfrost/block/src/main/java/com/bloxbean/cardano/yaci/store/blockfrost/block/controller/BFBlockController.java
@@ -73,7 +73,7 @@ public class BFBlockController {
     @GetMapping("epoch/{epoch_number}/slot/{slot_number}")
     @Operation(summary = "Block by epoch and slot", description = "Return block content for absolute slot constrained by epoch.")
     public BFBlockDTO getBlockByEpochAndSlot(@PathVariable("epoch_number") int epochNumber,
-                                             @PathVariable("slot_number") long slotNumber) {
+                                             @PathVariable("slot_number") int slotNumber) {
         return bfBlockService.getBlockByEpochAndSlot(epochNumber, slotNumber);
     }
